+++ src/modules/IdCardDetail/Intro.tsx
