--- src/modules/IdCardDetail/Intro/index.ts
@@ -1,1 +1,0 @@
-export * from './Intro';
