--- src/modules/IdCardDetail/KeywordContentCard/index.ts
@@ -1,1 +1,0 @@
-export * from './KeywordContentCard';
