+++ src/modules/IdCardDetail/index.ts
@@ -1,2 +1,3 @@
+export * from './IdCardDetail';
 export * from './Intro';
 export * from './KeywordContentCard';
